sh_api/requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
dj-database-url==0.5.0
Django==1.11.17
django-cors-headers==3.2.1
django-filter==2.2.0
django-allauth==0.30.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-environ==0.4.1
django-redis==4.7.0
django-rest-auth==0.9.1
django-sendgrid-v5==0.8.1
drfdocs
gunicorn==20.0.4
isort==4.3.21
ipython==5.2.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.0.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==5.0.1
wrapt==1.11.2
